Refactors explorer pool attachment.
